refactored dataset io and repo

+++ discover/infra/persist/object/dao.py
@@ -11,7 +11,7 @@
 # URL        : https://github.com/variancexplained/appvocai-discover                               #
 # ------------------------------------------------------------------------------------------------ #
 # Created    : Sunday September 22nd 2024 07:41:04 pm                                              #
-# Modified   : Wednesday December 25th 2024 10:45:11 pm                                            #
+# Modified   : Thursday December 26th 2024 02:09:16 am                                             #
 # ------------------------------------------------------------------------------------------------ #
 # License    : MIT License                                                                         #
 # Copyright  : (c) 2024 John James                                                                 #
@@ -34,7 +34,20 @@ from discover.infra.persist.object.base import DAO
 
 
 # ------------------------------------------------------------------------------------------------ #
+#                              SHELVE DATA ACCESS OBJECT                                           #
+# ------------------------------------------------------------------------------------------------ #
 class ShelveDAO(DAO):
+    """Data Access Object (DAO) implementation using Python's `shelve` module.
+
+    This class provides methods for managing assets in a persistent object database.
+    It supports CRUD operations and handles exceptions to ensure robust data access
+    and error reporting.
+
+    Args:
+        location (str): Directory where the database file is stored.
+        db_path (str): Name of the database file.
+        asset_type (AssetType): Type of asset managed by this DAO.
+    """
 
     def __init__(self, location: str, db_path: str, asset_type: AssetType):
         self._db_path = os.path.join(location, db_path)
@@ -44,13 +57,32 @@ class ShelveDAO(DAO):
 
     @property
     def count(self) -> int:
+        """Gets the count of all assets in the database.
+
+        Returns:
+            int: Number of assets in the database.
+        """
         return len(self.read_all)
 
     @property
     def size(self) -> int:
+        """Gets the size of the database file in bytes.
+
+        Returns:
+            int: Size of the database file in bytes.
+        """
         return os.path.getsize(self._db_path)
 
     def create(self, asset: Asset) -> None:
+        """Creates a new asset in the database.
+
+        Args:
+            asset (Asset): The asset to be added.
+
+        Raises:
+            ObjectDatabaseNotFoundError: If the database file is not found.
+            ObjectIOException: If an unknown exception occurs during creation.
+        """
         try:
             with shelve.open(self._db_path) as db:
                 db[asset.asset_id] = asset
@@ -64,6 +96,19 @@ class ShelveDAO(DAO):
             raise ObjectIOException(msg, e) from e
 
     def read(self, asset_id: str) -> Asset:
+        """Reads an asset by its ID from the database.
+
+        Args:
+            asset_id (str): The unique identifier of the asset to retrieve.
+
+        Returns:
+            Asset: The retrieved asset.
+
+        Raises:
+            ObjectNotFoundError: If the asset is not found in the database.
+            ObjectDatabaseNotFoundError: If the database file is not found.
+            ObjectIOException: If an unknown exception occurs during reading.
+        """
         try:
             with shelve.open(self._db_path) as db:
                 return db[asset_id]
@@ -81,7 +126,16 @@ class ShelveDAO(DAO):
             raise ObjectIOException(msg, e) from e
 
     def read_all(self) -> Dict[str, Asset]:
+        """Reads all assets from the database.
 
+        Returns:
+            Dict[str, Asset]: A dictionary of all assets, with keys as asset IDs
+            and values as Asset objects.
+
+        Raises:
+            ObjectDatabaseNotFoundError: If the database file is not found.
+            ObjectIOException: If an unknown exception occurs during reading.
+        """
         try:
             with shelve.open(self._db_path) as db:
                 return dict(db.items())
@@ -95,6 +149,18 @@ class ShelveDAO(DAO):
             raise ObjectIOException(msg, e) from e
 
     def exists(self, asset_id: str) -> bool:
+        """Checks if an asset exists in the database.
+
+        Args:
+            asset_id (str): The unique identifier of the asset to check.
+
+        Returns:
+            bool: True if the asset exists, False otherwise.
+
+        Raises:
+            ObjectDatabaseNotFoundError: If the database file is not found.
+            ObjectIOException: If an unknown exception occurs during the check.
+        """
         try:
             with shelve.open(self._db_path) as db:
                 return asset_id in db
@@ -108,6 +174,16 @@ class ShelveDAO(DAO):
             raise ObjectIOException(msg, e) from e
 
     def delete(self, asset_id: str) -> None:
+        """Deletes an asset by its ID from the database.
+
+        Args:
+            asset_id (str): The unique identifier of the asset to delete.
+
+        Raises:
+            ObjectNotFoundError: If the asset is not found in the database.
+            ObjectDatabaseNotFoundError: If the database file is not found.
+            ObjectIOException: If an unknown exception occurs during deletion.
+        """
         try:
             with shelve.open(self._db_path, writeback=True) as db:
                 del db[asset_id]
@@ -124,5 +200,26 @@ class ShelveDAO(DAO):
             self._logger.exception(msg)
             raise ObjectIOException(msg, e) from e
 
-    def reset(self) -> None:
-        shutil.rmtree(self._db_path)
+    def reset(self, verified: bool = False) -> None:
+        """Resets the database by deleting all its contents.
+
+        Args:
+            verified (bool): If True, performs the reset immediately. If False,
+                prompts the user for confirmation.
+
+        Logs:
+            Warning: Logs a warning if the reset is performed.
+            Info: Logs information if the reset operation is aborted.
+        """
+        if verified:
+            shutil.rmtree(self._db_path)
+            self._logger.warning(f"{self.__class__.__name__} has been reset.")
+        else:
+            proceed = input(
+                f"Resetting the {self.__class__.__name__} object database is irreversible. To proceed, type 'YES'."
+            )
+            if proceed == "YES":
+                shutil.rmtree(self._db_path)
+                self._logger.warning(f"{self.__class__.__name__} has been reset.")
+            else:
+                self._logger.info(f"{self.__class__.__name__} reset has been aborted.")
